Parse table of contents
There's an additional step for converting epub: Looking for a "toc.ncx"
which contains the table of contents and convert that to markdown. New
module NcxParser contains the logic for parsing table of contents and
converting to markdown.

convertFile has been moved to FileManager, it now takes the function
that converts to markdown as parameter since it could be ncxToMarkdown
or xhtmlToMarkdown.

Lib module has been renamed to XhtmlParser. LibTests module has been
renamed to XhtmlParserTests